Refactor location changed
